Rename voxel-related files
