+++ source/MRMesh/MRVolumeSegment.cpp
@@ -1,7 +1,6 @@
 #include "MRVolumeSegment.h"
 #ifndef MRMESH_NO_OPENVDB
 #include "MRVoxelPath.h"
-#include "MRSimpleVolume.h"
 #include "MRVolumeIndexer.h"
 #include "MRBox.h"
 #include "MRVoxelGraphCut.h"
